+++ src/components/quiz/ImportQuiz.tsx
@@ -9,8 +9,8 @@ import {
 } from "@chakra-ui/react";
 import { nanoid } from "nanoid";
 
-import { str2num } from "#/utils/commonFunctions";
 import db from "#/utils/db";
+import { str2num } from "#/utils/functions";
 
 const ImportQuiz: React.FC<{ setName: string }> = ({ setName }) => {
   const toast = useToast();
