+++ src/utils/computeScore/swedish10.ts
@@ -1,11 +1,11 @@
 import { GameDBProps, LogDBProps, WinPlayerProps } from "../types";
 
-import { numberSign } from "#/utils/commonFunctions";
 import {
   getInitialPlayersState,
   getSortedPlayerOrderList,
   indicator,
 } from "#/utils/computeScore";
+import { numberSign } from "#/utils/functions";
 
 // scoreをwrong ptとして利用
 const swedish10 = async (game: GameDBProps, gameLogList: LogDBProps[]) => {
